+++ runtime/src/lib.rs
@@ -7,6 +7,7 @@ extern crate lazy_static;
 pub mod account_info;
 pub mod account_overrides;
 pub mod account_rent_state;
+pub mod account_storage;
 pub mod accounts;
 pub mod accounts_background_service;
 pub mod accounts_cache;
