+++ runtime/src/serde_snapshot/tests.rs
@@ -3,8 +3,9 @@
 use {
     super::*,
     crate::{
+        account_storage::AccountStorageMap,
         accounts::{test_utils::create_test_accounts, Accounts},
-        accounts_db::{get_temp_accounts_paths, AccountShrinkThreshold, AccountStorageMap},
+        accounts_db::{get_temp_accounts_paths, AccountShrinkThreshold},
         accounts_hash::AccountsHash,
         append_vec::AppendVec,
         bank::{Bank, BankTestConfig},
